Create Memory manager
This project consists of the Process generator that simulates the arrival of processes. Every process is associated with random memory and the time that it spends in the system. The number of processes to be created will be mentioned by the user which can be denoted with ‘n’.  For allocating memory we use malloc() and for giving up the memory we use a free() system call. 
Even calculation for the time taken to allocate and free the memory is taken into consideration for estimating the total time taken by all the processes to execute. The same program ran multiple times and every time the time taken is evaluated and recorded for analysis. The process thread is started in the start_th function with PID as the process identifier and this process runs for a specified number of cycles to perform the task in the function process.